refactor(web): introduce src/domains structure + migrate announcements
Establish a domain-driven layout under apps/web/src/domains where each
product domain co-locates its components, server logic (queries/mutations/
permissions), and public barrel. App Router routes, shared UI, and
cross-cutting lib infra stay in place and import from the domain barrel.

Migrate the announcements domain as the first pilot:
- components/ <- components/announcements/*
- server/ <- lib/announcements.ts (visibility), create/update-announcement
- index.ts barrel + README; repoint all 4 importers at @/domains/announcements

Schemas remain in lib/schemas/content.ts (shared with events) as a tracked
follow-up. Typecheck passes.
