+++ apps/web/src/domains/announcements/components/AnnouncementCard.tsx
