+++ apps/web/src/domains/announcements/server/visibility.ts
@@ -1,6 +1,6 @@
 import type { SupabaseClient } from "@supabase/supabase-js";
 import type { Announcement, MembershipStatus } from "@/types/database";
-import type { OrgRole } from "./auth/role-utils";
+import type { OrgRole } from "@/lib/auth/role-utils";
 
 export type AnnouncementViewerContext = {
   role: OrgRole | null;
