+++ apps/web/src/lib/feed/load-feed-sidebar-data.ts
@@ -1,5 +1,5 @@
 import { createClient } from "@/lib/supabase/server";
-import { filterAnnouncementsForUserViaRpc } from "@/lib/announcements";
+import { filterAnnouncementsForUserViaRpc } from "@/domains/announcements";
 import type { OrgRole } from "@/lib/auth/role-utils";
 import type { Announcement, MembershipStatus } from "@/types/database";
 import type { ServerSupabase } from "@/lib/supabase/types";
